requirements.txt
python-telegram-bot ~= 13.8.1
requests~=2.26.0